RF: debian: Move dpkg-query line parser to support module
This is pure code movement.
